DHUD-367 Created Intertop parser DHUD-368 Add data to position provider, HUD panel service DHUD-370 Added Intertop site detection and settings DHUD-369 Create Intertop importer

+++ PokerHandHistoryParser/HandHistories.Objects/GameDescription/TournamentDescriptor.cs
@@ -38,10 +38,10 @@ namespace HandHistories.Objects.GameDescription
         public decimal TotalPrize { get; set; }
 
         [XmlElement]
-        public short FinishPosition { get; set; }
+        public int FinishPosition { get; set; }
 
         [XmlElement]
-        public short TotalPlayers { get; set; }
+        public int TotalPlayers { get; set; }
 
         [XmlElement]
         public DateTime StartDate { get; set; }
